[sosreport] move SoSOptions from sos.sosreport to sos
The options object now needs to be referenced in both the main
sosreport script, and the policy classes: move its definition
from the sos.sosreport module to sos to avoid recursive import
errors between sosreport and policy.
